[FIX]: Bug render facrory

+++ index.js
@@ -1,13 +1,13 @@
 import "./show-factory.js"
 import "./buttons.js"
 import './show-form.js' 
-import "./crud.js"
+import './crud.js'
 import { getAllFactoryes } from "./crud.js"
-
-let data = await getAllFactoryes()
+// Данные из монго записываются в localStorage
 function store(factory){
+    if(factory){
         localStorage.clear()
         localStorage.setItem('factoryes', `${JSON.stringify(factory)}`)
+    }
 }
-store(data)
-
+store(await getAllFactoryes())
